requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.8.0
exceptiongroup==1.3.1
fastapi==0.124.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.55.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
inngest==0.5.13
jcs==0.2.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.1.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.0
langchain-openai==1.1.3
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
llama-index-core==0.14.10
llama-index-instrumentation==0.4.2
llama-index-readers-file==0.5.5
llama-index-workflows==2.11.5
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
ollama==0.6.1
openai==2.11.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.2.3
pillow==12.0.0
platformdirs==4.5.1
portalocker==3.2.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pymongo==4.15.5
pyparsing==3.2.5
pypdf==6.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
streamlit==1.52.1
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.3
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
